+++ sphinx/util/docutils.py
@@ -566,6 +566,13 @@ class SphinxTranslator(nodes.NodeVisitor):
         self.builder = builder
         self.config = builder.config
         self.settings = document.settings
+        #: Indicate whether this translator is acting on a partial document
+        #: or not. This is typically useful when, in some visitor methods,
+        #: it is known that the method could be called from either a regular
+        #: document or a partial one which may not necessarily contain nodes
+        #: that were transformed (e.g., titles nodes stored by the environment
+        #: are not processed by Sphinx transformations).
+        self.in_partial_document = is_partial_document(document)
 
     def dispatch_visit(self, node: Node) -> None:
         """
@@ -633,3 +640,21 @@ def new_document(source_path: str, settings: Any = None) -> nodes.document:
     document = addnodes.document(settings, reporter, source=source_path)
     document.note_source(source_path, -1)
     return document
+
+
+def new_partial_document(settings: Any = None) -> nodes.document:
+    """Return a new empty partial document.
+
+    A partial document is a document with minimal capabilities and which
+    contains nodes that may not have been transformed by the docutils or
+    Sphinx transformations.
+
+    This function is typically used to render partial fragments of a doctree
+    using non-transformed information, e.g., the HTML ``<title>`` tag.
+    """
+    return new_document('<partial node>', settings)
+
+
+def is_partial_document(document: nodes.document):
+    """Detect whether *document* is a partial document or not."""
+    return document.get('source') == '<partial node>'